Migrated `BlockedUser`, `ChangePw`, `ChannelCreate` to atomic templates
