+++ client/src/components/pages/ChannelCreateContainer.tsx
@@ -1,13 +1,5 @@
-import {
-  Animated,
-  FlatList,
-  Image,
-  ScrollView,
-  Text,
-  TouchableOpacity,
-  View,
-} from 'react-native';
-import {Channel, User, UserConnection, UserEdge} from '../../types/graphql';
+import {Animated, Text, TouchableOpacity, View} from 'react-native';
+import {Channel, User, UserConnection} from '../../types/graphql';
 import {
   ChannelCreateFriendsPaginationQuery,
   ChannelCreateFriendsPaginationQueryVariables,
@@ -16,8 +8,6 @@ import {
   ChannelFindOrCreatePrivateChannelMutation,
   ChannelFindOrCreatePrivateChannelMutationResponse,
 } from '../../__generated__/ChannelFindOrCreatePrivateChannelMutation.graphql';
-import {IC_CIRCLE_X, IC_NO_IMAGE} from '../../utils/Icons';
-import {LoadingIndicator, useTheme} from 'dooboo-ui';
 import React, {FC, ReactElement, Suspense, useMemo, useState} from 'react';
 import {
   graphql,
@@ -26,20 +16,18 @@ import {
   usePaginationFragment,
 } from 'react-relay/hooks';
 
+import ChannelCreateTemp from '../templates/ChannelCreateTemp';
 import {ChannelCreate_friends$key} from '../../__generated__/ChannelCreate_friends.graphql';
-import ErroView from '../UI/molecules/ErrorView';
 import {FriendsQuery} from '../../__generated__/FriendsQuery.graphql';
+import {LoadingIndicator} from 'dooboo-ui';
 import {MainStackNavigationProps} from '../navigations/MainStackNavigator';
 import SearchTextInput from '../UI/atoms/SearchTextInput';
-import UserListItem from '../UI/molecules/UserListItem';
 import {findOrCreatePrivateChannel} from '../../relay/queries/Channel';
 import {friendsQuery} from '../../relay/queries/Friend';
 import {getString} from '../../../STRINGS';
-import produce from 'immer';
 import {showAlertForError} from '../../utils/common';
 import styled from 'styled-components/native';
 import useDebounce from '../../hooks/useDebounce';
-import {useNavigation} from '@react-navigation/native';
 
 const ITEM_CNT = 20;
 
@@ -47,13 +35,6 @@ const Container = styled.SafeAreaView`
   flex: 1;
   background-color: ${({theme}) => theme.backgroundDark};
   flex-direction: column;
-`;
-
-const FriendThumbView = styled.View`
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  width: 70px;
 `;
 
 interface ChannelCreate extends User {
@@ -110,6 +91,8 @@ type FriendsFragmentProps = {
 
 const FriendsFragment: FC<FriendsFragmentProps> = ({
   friend,
+  scrollY,
+  searchArgs,
   selectedUsers,
   setSelectedUsers,
 }) => {
@@ -126,145 +109,14 @@ const FriendsFragment: FC<FriendsFragmentProps> = ({
     );
   }, [data]);
 
-  const {theme} = useTheme();
   // const [friends, setFriends] = useState<Friend[]>(fakeFriends);
 
-  const navigation = useNavigation();
-
-  const removeFriend = (friendArg: User): void => {
-    const nextState = produce(selectedUsers, (draftState) => {
-      const index = selectedUsers.findIndex((v) => v.id === friendArg.id);
-
-      draftState.splice(index, 1);
-    });
-
-    setSelectedUsers(nextState);
-  };
-
-  const renderFriendThumbnail = (
-    friendArg: User,
-    index: number,
-  ): ReactElement => {
-    return (
-      <FriendThumbView key={friendArg.id}>
-        <View
-          style={{
-            marginTop: 12,
-            marginRight: 16,
-            marginBottom: 6,
-            justifyContent: 'center',
-            alignItems: 'center',
-          }}>
-          <Image
-            style={{
-              width: 60,
-              height: 60,
-            }}
-            source={
-              friendArg.thumbURL ? {uri: friendArg.thumbURL} : IC_NO_IMAGE
-            }
-          />
-          <Text
-            numberOfLines={1}
-            style={{
-              fontSize: 12,
-              marginTop: 4,
-              color: theme.text,
-            }}>
-            {friendArg.nickname || friendArg.name}
-          </Text>
-        </View>
-        <TouchableOpacity
-          testID={`remove-${index}`}
-          style={{
-            position: 'absolute',
-            top: 0,
-            right: 0,
-          }}
-          onPress={(): void => removeFriend(friendArg)}>
-          <Image source={IC_CIRCLE_X} style={{width: 32, height: 32}} />
-        </TouchableOpacity>
-      </FriendThumbView>
-    );
-  };
-
-  const renderItem = ({
-    item,
-    index,
-  }: {
-    item: UserEdge;
-    index: number;
-  }): ReactElement => {
-    return (
-      <UserListItem
-        testID={`userlist_${index}`}
-        showCheckBox
-        checked={selectedUsers.includes(item?.node as User)}
-        // @ts-ignore
-        user={item.node}
-        onPress={(): void => {
-          if (selectedUsers.includes(item?.node as User)) {
-            const nextState = produce(selectedUsers, (draftState) => {
-              const selectedUserIndex = selectedUsers.findIndex(
-                (v) => v.id === item?.node?.id,
-              );
-
-              draftState.splice(selectedUserIndex, 1);
-            });
-
-            setSelectedUsers(nextState);
-
-            return;
-          }
-
-          const nextState = produce(selectedUsers, (draftState) => {
-            draftState.push(item.node as User);
-          });
-
-          setSelectedUsers(nextState);
-        }}
-      />
-    );
-  };
-
   return (
-    <Container>
-      <FlatList
-        style={{alignSelf: 'stretch'}}
-        contentContainerStyle={
-          friendEdges.length === 0
-            ? {
-                flex: 1,
-                alignSelf: 'stretch',
-                flexDirection: 'column',
-              }
-            : null
-        }
-        keyExtractor={(_, index): string => index.toString()}
-        data={friendEdges as UserEdge[]}
-        renderItem={renderItem}
-        ListHeaderComponent={(): ReactElement => {
-          return (
-            <ScrollView
-              style={{paddingHorizontal: 24, marginBottom: 12}}
-              horizontal>
-              {selectedUsers.map((selectedUser, i) =>
-                renderFriendThumbnail(selectedUser, i),
-              )}
-            </ScrollView>
-          );
-        }}
-        ListEmptyComponent={
-          <ErroView
-            testID="btn-error"
-            title={getString('NO_FRIEND')}
-            body={getString('NO_FRIEND_BODY')}
-            buttonText={getString('GO_BACK')}
-            onButtonPressed={(): void => navigation.goBack()}
-          />
-        }
-      />
-    </Container>
+    <ChannelCreateTemp
+      friendEdges={friendEdges}
+      selectedUsers={selectedUsers}
+      setSelectedUsers={setSelectedUsers}
+    />
   );
 };
 
@@ -306,6 +158,7 @@ const ChannelCreate: FC<ChannelCreateProps> = (props) => {
   const {navigation} = props;
   const [searchText, setSearchText] = useState<string>('');
   const [selectedUsers, setSelectedUsers] = useState<User[]>([]);
+
   const debouncedText = useDebounce(searchText, 500);
   const scrollY = new Animated.Value(0);
